correcting PR review, moving matches logic to a new matches controller
