+++ app/controllers/games.js
@@ -1,5 +1,4 @@
 const gameService = require('../services/games'),
-  matchesService = require('../services/matches'),
   errors = require('../errors');
 
 const gameValidation = (game) => {
@@ -8,15 +7,6 @@ const gameValidation = (game) => {
   }
   if (!game.name) {
     throw errors.validationError('Name is missing');
-  }
-};
-
-const matchValidation = (match) => {
-  if (!match.game_id) {
-    throw errors.validationError('Game_id is missing');
-  }
-  if (!match.assertions) {
-    throw errors.validationError('Assertions value is missing');
   }
 };
 
@@ -41,16 +31,3 @@ exports.list = (req, res, next) => {
   });
 };
 
-exports.createMatch = (req, res, next) => {
-  const match = {
-    user_id: req.user.id,
-    game_id: req.params.game_id,
-    assertions: req.body.assertions
-  };
-  matchValidation(match);
-  matchesService.create(match).then((createdMatch) => {
-    res.jsonp(createdMatch);
-  }).catch((err) => {
-    next(err);
-  });
-};
