+++ docs/admin.html
@@ -1,0 +1,24 @@
+<html><head><link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/bootstrap/3.3.6/css/bootstrap.min.css" /><script src="https://code.jquery.com/jquery-1.12.2.min.js"></script><script src="https://maxcdn.bootstrapcdn.com/bootstrap/3.3.6/js/bootstrap.min.js"></script><style> pre { width: 95%; } </style></head><body><div class="container"><div class="jumbotron"><h1>Admin</h1></div><div class="panel panel-default"><div class="panel-heading" role="tab" id="heading-admin-users--POST"><h4 class="panel-title"><a role="button" data-toggle="collapse" data-parent="#accordion" aria-expanded="true" aria-controls="collapse-admin-users--POST" href="#collapse-admin-users--POST">POST /admin/users/</a></h4></div><div id="collapse-admin-users--POST" class="panel-collapse collapse" role="tabpanel" aria-labelledby="heading-admin-users--POST"><div class="panel-body"><h4>Request Data</h4><div class="container"><h5>Headers</h5><pre>{
+  "authorization": "eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJhdXRoZW50aWNhdGlvbkNvZGUiOiIkMmEkMTAkQXZXSGFPenNISVdpdzJTdkxzR081TyIsIm1heGltdW1Vc2VmdWxEYXRlIjoiMjAxNy0wOC0yNlQxOTozMzoxMS40MTZaIiwiZXhwaXJhdGlvbkRhdGUiOiIyMDE3LTA3LTI5VDE5OjMzOjExLjQxN1oiLCJpZCI6MiwicmVuZXdJZCI6IiQyYSQxMCRyWlpLMnU0SHZuS2JzOUJFUWJqMVBPIn0.40vM_Cl7A_-Ao1sFCWktQPzwUrfsKfF6LPNXDSElerc",
+  "content-type": "application/json",
+  "content-length": 100
+}</pre><h5>Body</h5><pre>{
+  "firstName": "xxxxxx",
+  "lastName": "zzzzzzzzz",
+  "password": "12345678910",
+  "email": "123456@wolox.com.ar"
+}</pre></div></div><div class="panel-body"><h4>Response Data</h4><div class="container"><h5>Status: 201</h5><h5>Headers</h5><pre>{
+  "authorization": "eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJhdXRoZW50aWNhdGlvbkNvZGUiOiIkMmEkMTAkQXZXSGFPenNISVdpdzJTdkxzR081TyIsIm1heGltdW1Vc2VmdWxEYXRlIjoiMjAxNy0wOC0yNlQxOTozMzoxMS40MTZaIiwiZXhwaXJhdGlvbkRhdGUiOiIyMDE3LTA3LTI5VDE5OjMzOjExLjQxN1oiLCJpZCI6MiwicmVuZXdJZCI6IiQyYSQxMCRyWlpLMnU0SHZuS2JzOUJFUWJqMVBPIn0.40vM_Cl7A_-Ao1sFCWktQPzwUrfsKfF6LPNXDSElerc",
+  "content-type": "application/json",
+  "content-length": 100
+}</pre><h5>Body</h5><pre>{
+  "id": 5,
+  "firstName": "xxxxxx",
+  "lastName": "zzzzzzzzz",
+  "password": "$2a$10$cD7oQ604yF.cJP/uyYfrHeShQkrED2ZB7Xw8xVKMjvdLlJpPs/eOa",
+  "email": "123456@wolox.com.ar",
+  "isAdmin": true,
+  "authenticationCode": "$2a$10$mDT2jCZzTb33b6RwM.IHzO",
+  "updatedAt": "2017-07-27T19:33:11.496Z",
+  "createdAt": "2017-07-27T19:33:11.496Z"
+}</pre></div></div></div></div></div></body></html>
